+++ config/dev.exs
@@ -8,4 +8,5 @@ config :eventstore, EventStore.Storage,
   password: "postgres",
   database: "eventstore_dev",
   hostname: "localhost",
-  pool_size: 10
+  pool_size: 10,
+  extensions: [{Postgrex.Extensions.Calendar, []}]